docs: Added README file
Included a note regarding the fact that there no clear
understanding on how to point a specific fusion360 script or App
to a specific environment.

This means that the app would probabley have to act as a facade or
as an adaptor to an API entry to other services used for analysis.

TODO: Understand how environments work for fusion
TODO: Ubderstand how to use docker compose in deveopment